Bug 1654270 - Properly restart after a partial match if we reach the end of the document. r=jfkthame
This was kind of a pre-existing bug, though not surfaced because we
didn't have conditionally-appearing text nodes I guess.

If we are doing a match in a non-anonymous subtree, we'd skip the
anonymous text (that's alright). But if we actually reach the end of the
document (as it is the case on the test that starts failing with the
other patch in this bug), we'd just early-return, rather than restarting
where we left off, which is not fine and means that potential matches in
anonymous subtrees that were skipped won't show up.

This fixes the bug... Eventually we should probably come up with a
better structure than this gigantic while (true), probably moving all
these variables into their own struct, but today is not that day :)

Differential Revision: https://phabricator.services.mozilla.com/D84425

UltraBlame original commit: aa337bc4d6693ad1d25236229eb982ebe468e138