Bug 1648344 Part 2: Cleanup Document to remove NoValidContent viewport type. r=hiro
The NoValidContent type was added to detect cases where we needed to apply
default width the viewport. That approach is no longer needed, and this
enum can be removed, as well as the code that sets and checks that enum.

Differential Revision: https://phabricator.services.mozilla.com/D84447

UltraBlame original commit: 5900b99f8c02d90e2adbd87aa6be8d9a06a08541