Bug 1653292: Update range of accepted numbers for fractionalSecondDigits. r=jwalden
`DefaultNumberOption()` was updated to allow `fallback` to be `undefined` and
the code was slightly reordered to match the current spec text.

"fractionalSecondDigits" can now also be handled in the common loop to detect
if it is present when either "dateStyle" or "timeStyle" is used.

Differential Revision: https://phabricator.services.mozilla.com/D83948

UltraBlame original commit: b3837fc631658eb4dd4261fe086279984f6311ae