Bug 1654947 - Don't clobber RHS register for shift instructions. r=iain
The masking was unnecessary anyway on x86/x64/arm64 because the CPU does it (and
we are already relying on that in Ion). For arm32 handle this in the flexible*
MacroAssembler implementation by using the scratch register.

Differential Revision: https://phabricator.services.mozilla.com/D84787

UltraBlame original commit: f5c234cd9add0d84481a1bd17b6519739829f8e6