Bug 1654769 - Make clientRect offsets relative to the primary frame of the element. r=mats
We're returning offsets relative to the scroll target frame, which for
inputs and such is not the primary frame, and thus they never have
border or padding or what not.

Offsets should be relative to the primary frame however, so move the
rect appropriately.

Differential Revision: https://phabricator.services.mozilla.com/D84919

UltraBlame original commit: 0ce4c53c6abb4e5325af3cd0505ec943d99e74a9