Bug 1649718 - Introduce fuzzing intercession points in feature switches.  r=bbouvier
The fuzzing intercession points allow us to switch off a feature if fuzzing is
selected and only compilers that don't support the feature are enabled.
To be conservative, I limit it to cases where only a single compiler is
selected.

This does introduce another point in the code where we have to maintain feature
selection, but as the number of unsupported features goes down, this should
remain maintainable, I think.

Differential Revision: https://phabricator.services.mozilla.com/D81840

UltraBlame original commit: 4848ee762ec8fa5fe448f8e6d5ffeec1f814bcbd