Bug 1649718 - Do not panic when fuzzing results in nonsensical flags. r=bbouvier
Fuzzing can get us into a situation where we have only Ion (or
Cranelift) but debugging is also enabled.  Guards higher up in the
stack would normally prevent us from getting as far as the wasm
compiler, but the fuzzing intercession point introduced in the
previous patch disables the sniffing that causes the presence of the
debugger to disable Ion; thus we come in here with debug && ion.

Since it's fine to turn this from a release assert into a run-time
error, in the way where we error out below in the case when no
compilers are defined, then let's do so, instead of making things even
more complicated.

Depends on D81840

Differential Revision: https://phabricator.services.mozilla.com/D84990

UltraBlame original commit: 67f9e60ed1db9f1eda4cbf5f2016686f57a3f087