Bug 1651799 - Add test to ensure experimental features are not released. r=bbouvier
This commit adds a test that tracks all experimental wasm features and asserts
that if they are enabled, then we must be in nightly. Each feature is also
given a small feature detection module that will fail to compile when it's
not enabled for extra protection.

Differential Revision: https://phabricator.services.mozilla.com/D83424

UltraBlame original commit: f3792719ddb6c57476d9503694c48b27637f50cc