Bug 1653667 don't attempt to promote GraphRunner thread on Linux r=padenot
This usually fails due to sandbox policy.

When the sandbox is disabled, the realtime promotion can result in SIGKILL
if/when script runs too long.  This can be re-introduced when thread demotion
is moved to a thread that is not blocked on script, which may be part of
the solution to https://bugzilla.mozilla.org/show_bug.cgi?id=1628198.

Depends on D84941

Differential Revision: https://phabricator.services.mozilla.com/D84942

UltraBlame original commit: 60e3fc093c792f081cfa3dd0d0aeba02b58ddfe5