Bug 1635662 - Manage the ecosystem client ID in telemetry. r=chutten
This commit adds a new ecosystem client ID to `ClientID.jsm`. It's
handled the same way as a regular telemetry client ID, in that it's a
random UUID that's persisted to disk, and reset together with the
client ID. Unlike the main client ID, we don't cache it in prefs.

Additionally:

* `setClientID` has been renamed to `setCanaryClientIDs`, and
  automatically sets the main and ecosystem client IDs to the canary
  (known) ID. It's no longer possible to set either client ID to a
  specific value, but only test code did this.
* `resetClientID` has been renamed to `removeClientIDs`, and clears
  both client IDs. It no longer returns the new ID; instead, consumers
  should call `getClientID` or `getEcosystemClientID` after.
* The ecosystem client ID is included in the `deletion-ping`.

Differential Revision: https://phabricator.services.mozilla.com/D84233

UltraBlame original commit: cf7bb4a22eb93ba0cc9f80515cdb8a6787c62294