Bug 1655608 part 3: Fix ReflowOutput/ReflowInput mismatch in nsPageSequenceFrame. r=TYLin
For a given ::Reflow call, we have an expectation that the ReflowInput and
ReflowOutput should be "matching".  In particular: we should use the
ReflowInput to initialize the ReflowOutput, and we should pass both of these to
the frame's Reflow method.

In current mozilla-central (prior to this patch), nsPageSequenceFrame violates
this expectation when it prepares to reflow its children, in the following
ways:
1) It passes in the wrong ReflowInput as an arg when constructing a
   ReflowOutput struct for its children. (It passes *its own* ReflowInput arg,
   rather than the ReflowInput for the child that it's about to reflow.)
2) It *reuses* that single ReflowOutput structure *for all of its children*,
   which means that sizing information from one child will end up improperly
   influencing later children by "leaking through".

This results in overflow areas that are artificially too large on nsPageFrames,
because ReflowOutput populates its overflow areas via "union()" operations. So
the improperly-reused ReflowOutput produces a strictly larger overflow area
on each subsequent page (even if the page doesn't actually have any overflow).

This patch fixes these problems by creating a new ReflowOutput for each child,
and initializing it with the correct ReflowInput instance.

Notably, this patch *is* a behavior-change (in the direction of correctness).
In particular, we won't have artificially-inflated overflow areas on later
nsPageFrame instances anymore, and so now some of our automated testcases
trigger assertion-failures that they were getting around before. These all
seem to be versions of https://bugzilla.mozilla.org/show_bug.cgi?id=1655630
which, "fortunately", is already reproducible (under certain paper sizes)
even in a build from current mozilla-central, without this here patch.

Differential Revision: https://phabricator.services.mozilla.com/D85098

UltraBlame original commit: cbb878f5dc662491a20b7dd36fcd11a1727c1173