Bug 1654516 part 1 - Optimize _IsConstructing intrinsic in CacheIR and Warp. r=iain
This is a pretty big performance improvement for bound functions in Baseline.

The transpiler part was non-trivial becuase the transpiler has to know whether
we're inlining and constructing. Instead of adding yet another argument to
TranspileCacheIRToMIR, this passes the WarpBuilder so we can use its inlineCallInfo.

Differential Revision: https://phabricator.services.mozilla.com/D84517

UltraBlame original commit: 1152d7b452a994442b8e185956101903aee3c922