Bug 1617002 - Rework GTK opaque region handling. r=stransky
* always update the opaque region if the toplevel window allocation
changes - this fixes opaque regions on X11 with WR and OGL renderers,
as GTK resets the opaque region in this case. The previous solution
was modeled similar to what's done on Windows - we have to follow GTK
though.
 * use `gdk_window_set_opaque_region` also for Wayland - this fixes
occasional glitches where the opaque region was not properly updated.
Let GTK handle that instead, as it has better control of the internals.
 * remove a failed attempt to work around bug 1615098 from
MozContainerWayland
 * repurpose `widget.wayland.use-opaque-region` for subsurface opaque
regions and disable it by default. We want to enable it again eventually
and not all Wayland compositors suffer from bug 1615098
 * some cleanups - some optimizations should not be needed anymore now

Unfortunately this does not cover the classic title bar yet - that will
need to be done in a follow-up.

Differential Revision: https://phabricator.services.mozilla.com/D84024

UltraBlame original commit: 16bf991df0003caa9a5db83448373fb4f0bcd561