Bug 1655031 - Add a google image search performance test. r=acreskey
This patch adds a test for scrolling on the google image search page for kittens. It measure the images loaded per second as well as the amount of time it takes for a clicked image to open and load.

Differential Revision: https://phabricator.services.mozilla.com/D84927

UltraBlame original commit: 547f88256a8c6274d9418b54cd708c8ed0d684fc