Bug 1654994 - Allow upload_symbols to handle .tar.zst symbols archives. r=firefox-build-system-reviewers,rstewart,mhentges
Because the symbols server can only handle the .zip format for the
symbols archives, we make upload_symbols itself munge what it takes from
builds if they are .tar.zst, so that the symbols server is happy.

Eventually, the symbols server will be able to deal with those archives
on its own.

It is worth noting that this conversion is actually taking more time
than builds spend for the same archive because nothing is parallelized,
while in builds the compression of e.g. libxul.so and libxul-gtest
happen in parallel.

However, overall, we'll still gain from switching to this new code path
because builds currently always compress the debug info, even when it's
not going to be uploaded to the symbols server.

Differential Revision: https://phabricator.services.mozilla.com/D85250

UltraBlame original commit: c31ebb9aae4a66dddd3117c12e8dbefdcb3f9c1a