Bug 1655393 - Add WPT for `InputEvent.getTargetRanges()` r=smaug
There are no automated tests for `InputEvent.getTargetRanges()` because it
is set only when `beforeinput` event, but it's defined as not dispatched by
`document.execCommand`.

However, we can synthesize `beforeinput` event with test driver.

On the other hand, the definition in Input Events spec is not clear.
Therefore, most of the tests won't be passed on any browsers for now.
There are some spec issues which I filed:
* https://github.com/w3c/input-events/issues/112
* https://github.com/w3c/input-events/issues/113
* https://github.com/w3c/input-events/issues/114

These new test must be useful when browser vendors discuss the issues.

Differential Revision: https://phabricator.services.mozilla.com/D85527

UltraBlame original commit: 12f21ad909371384939bb38d65cf1dbc797b5bf9